Sign in screen (#10)
* Routing

Added a SignInSignUpScreen to let you know if that is showen if the user isn't signIn. There's also the SignIn page but that has to be validated. The Router now points to the SignInSignUpScreen.

* Update SignInSignUpScreen.kt

commented something that wasn't used

* SignInScreen Refactoring

I've put the Information in a column and try to see what it return after signIn with firebase

* Validation

logging the email and the password!

* Trying to figure out the signIn

* Recover all code

* Update UserSignUpScreen.kt

Slight change

* SignInScreen

loging of the different state

* fixed ktlint issue 'standard:function-start-of-body-spacing'

---------

Co-authored-by: JeanRose-JAC <98351000+JeanRose-JAC@users.noreply.github.com>